+++ src/main/scala/app/controllers/MessagingController.scala
@@ -1,36 +1,12 @@
-package app.business.routes
+package app.controllers
 
-import app.model._
 import app.model.Message._
-import sttp.model.{CookieValueWithMeta, StatusCode}
-import sttp.tapir._
-import sttp.tapir.docs.openapi._
-import sttp.tapir.json.circe._
-import sttp.tapir.openapi.OpenAPI
-import sttp.tapir.openapi.circe.yaml._
+import app.model._
+import sttp.model.StatusCode
+import sttp.tapir.json.circe.jsonBody
+import sttp.tapir.{auth, cookie, endpoint, oneOf, statusCode, statusMapping, Endpoint}
 
-/**
-  * Tapir routes description with OpenAPI spec generation
-  * */
-object RoutesDescription extends JsonCodecs {
-
-  val health: Endpoint[Unit, Unit, StatusCode, Nothing] =
-    endpoint.get
-      .in("health")    // Path
-      .out(statusCode) // Output type
-      .tag("Util")
-
-  val signIn: Endpoint[Authorize, StatusCode, CookieValueWithMeta, Nothing] =
-    endpoint.post
-      .in("auth")
-      .in(jsonBody[Authorize])
-      .out(setCookie("sessionid"))
-      .errorOut(statusCode)
-      .tag("Messenger")
-      .summary("Sign in with user id and password")
-      .errorOut(
-        statusCode(StatusCode.Forbidden).description("Invalid user ID or password")
-      )
+object MessagingController extends JsonCodecs {
 
   val send: Endpoint[(Option[String], IncomingTextMessage), ErrorInfo, StatusCode, Nothing] =
     endpoint.post
@@ -97,32 +73,4 @@ object RoutesDescription extends JsonCodecs {
       .errorOut(statusCode)
       .tag("Messenger")
       .summary("Get all messages from specified time from server")
-
-  val authTest: Endpoint[Option[String], StatusCode, StatusCode, Nothing] =
-    endpoint.get
-      .in("authTest")
-      .in(auth.apiKey(cookie[Option[String]]("sessionid")))
-      .out(statusCode)
-      .errorOut(statusCode)
-      .tag("Util")
-      .summary("Test if cookie is valid")
-
-  /**
-    * List of endpoints for generating OpenAPI doc
-    * */
-  private val openApi: OpenAPI = List(
-    health,
-    send,
-    addToConversation,
-    conversations,
-    sync,
-    signIn,
-    authTest
-  ).toOpenAPI(
-    "Scala Tapir Messenger",
-    "0.0.1"
-  )
-
-  /** One only need to create with this openApiYml for each server type and it just works */
-  val openApiYml: String = openApi.toYaml
 }
